requirements.txt
flask>=3.0
opencv-python>=4.8
Pillow>=10.0
numpy>=1.24
face_recognition>=1.2
dlib>=19.24
reverse_geocoder>=1.5
